requirements.txt
tqdm
numpy
librosa
soundfile
google-cloud-texttospeech
